requirements.txt
altair==5.5.0
altgraph==0.17.4
asgiref==3.10.0
attrs==25.4.0
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==3.2.20
django-widget-tweaks==1.5.0
djongo==1.3.6
Flask==3.1.2
fonttools==4.59.2
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
gunicorn==23.0.0
idna==3.11
imageio==2.37.0
imbalanced-learn==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaggle==1.7.4.5
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
names==0.3.0
narwhals==2.9.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pefile==2023.2.7
pillow==11.3.0
plotly==6.4.0
protobuf==6.33.0
pyarrow==21.0.0
pydeck==0.9.1
pygame==2.6.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pymongo==3.11.4
PyMySQL==1.1.2
pyngrok==7.4.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sqlparse==0.2.4
statsmodels==0.14.5
streamlit==1.50.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
webencodings==0.5.1
Werkzeug==3.1.3
